+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 ### Updates & Improvements
-- Updated Fabric to version 1.21.9–10
-- Dropped support for Forge and NeoForge *(may reintroduce NeoForge support in the future)*
+- Updated Fabric to version 1.21.11